consolidation of OtlpOptions
