--- domains/certificates/Query.API/API/Configurations/OtlpOptions.cs
@@ -1,12 +1,0 @@
-using System;
-using System.ComponentModel.DataAnnotations;
-
-namespace API.Configurations;
-
-public class OtlpOptions
-{
-    public const string Prefix = "Otlp";
-
-    [Required]
-    public Uri ReceiverEndpoint { get; set; } = null!;
-}
