--- domains/measurements/Measurements.API/API/Options/OtlpOptions.cs
@@ -1,12 +1,0 @@
-using System;
-using System.ComponentModel.DataAnnotations;
-
-namespace API.Options;
-
-public class OtlpOptions
-{
-    public const string Prefix = "Otlp";
-
-    [Required]
-    public Uri ReceiverEndpoint { get; set; } = null!;
-}
