+++ domains/transfer/Transfer.API/API/Program.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using System.Net.Http;
 using System.Text.Json.Serialization;
 using API.Cvr;
-using API.Shared.Options;
 using API.Transfer;
 using API.Transfer.Api.Clients;
 using API.UnitOfWork;
@@ -12,6 +11,7 @@ using EnergyOrigin.ActivityLog;
 using EnergyOrigin.Setup;
 using EnergyOrigin.Setup.Exceptions.Middleware;
 using EnergyOrigin.Setup.Migrations;
+using EnergyOrigin.Setup.OpenTelemetry;
 using EnergyOrigin.TokenValidation.b2c;
 using FluentValidation;
 using Microsoft.AspNetCore.Builder;
