--- domains/transfer/TransferAgreementAutomation/Worker/Options/OtlpOptions.cs
@@ -1,12 +1,0 @@
-using System;
-using System.ComponentModel.DataAnnotations;
-
-namespace TransferAgreementAutomation.Worker.Options;
-
-public class OtlpOptions
-{
-    public const string Prefix = "Otlp";
-
-    [Required]
-    public Uri ReceiverEndpoint { get; set; } = null!;
-}
