+++ domains/transfer/TransferAgreementAutomation/Worker/Program.cs
@@ -1,6 +1,7 @@
 using System;
 using DataContext;
 using EnergyOrigin.Setup;
+using EnergyOrigin.Setup.OpenTelemetry;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
